// webview-ui/src/thor/model/PtgRef.ts
export type PtgRef  = {
    /**
     * 
     * @type {string}
     * @memberof PtgRef
     */
    sheetId?: string;
    /**
     * 
     * @type {string}
     * @memberof PtgRef
     */
    formulaId?: string;
    /**
     * 
     * @type {string}
     * @memberof PtgRef
     */
    reference?: string;
    /**
     * 
     * @type {string}
     * @memberof PtgRef
     */
    type?: PtgRefTypeEnum;
    /**
     * Unique identifier for object in the system
     * @type {string}
     * @memberof PtgRef
     */
    id?: string;
    /**
     * UUID of owner of the object in the system
     * @type {string}
     * @memberof PtgRef
     */
    ownerId?: string;
    /**
     * Date of object creation
     * @type {Date}
     * @memberof PtgRef
     */
    createdDate?: Date;
    /**
     * Data, including hash of the key(s) used to encrypt this record.
     * @type {string}
     * @memberof PtgRef
     */
    keyHash?: string;
    /**
     * Last user to access object
     * @type {string}
     * @memberof PtgRef
     */
    lastAccessedById?: string;
    /**
     * Timestamp of last access of object
     * @type {Date}
     * @memberof PtgRef
     */
    lastAccessedDate?: Date;
    /**
     * Unique identifier for user who last modifed the object in the system
     * @type {string}
     * @memberof PtgRef
     */
    lastModifiedById?: string;
    /**
     * Date of last object modification
     * @type {Date}
     * @memberof PtgRef
     */
    lastModifiedDate?: Date;
}
// ...
export enum PtgRefTypeEnum {
    PTGADD = 'ptgadd',
    PTGSUB = 'ptgsub',
    PTGMUL = 'ptgmul',
    PTGDIV = 'ptgdiv'
}